requirements.txt
python-telegram-bot
python-dotenv
freegames
art
antigravity
emoji
Pyjokes
faker
pillow
